feat: enhance login box on endorsements page
